requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==6.1.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.2.1
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
